+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 - `TRANSPORT` and `TRANSPORTALL` input options; these were obsolete
 - `LUMP` input option; this was obsolete
+- `DEFRAD`, `SETRAD`, `INITIALIZE`, `XGRID`, `YGRID`, `ZGRID`, `WRITE_OPT`, `RUN`, `USE`, `USES`; these were obsolete
 
 ## [3.0.2] - 2023-06-02
 ### Added